git commit -m "Fix record_id retrieval and storage in stash
In this session, we identified and fixed an issue with the retrieval and storage of the record_id in the stash. The record_id was not being correctly retrieved from the URL parameters and stored in the stash. We updated the log_form subroutine to correctly retrieve the record_id from the URL parameters and store it in the stash. This allows the record_id to be correctly passed to the form and used in the creation of new log entries."